move asset stanza gen to create_
