+++ slk/config/templates/sidechain.jinja
@@ -101,6 +101,8 @@ mainchain_account={{ mainchain_door_account }}
 mainchain_ip={{ mainchain_ip }}
 mainchain_port_ws={{ mainchain_port_ws }}
 
+{{ asset_stanzas }}
+
 {{ sidechain_stanza }}
 
 [features]
