+++ slk/create_config_files.py
@@ -17,15 +17,16 @@ the environment variable RIPPLED_MAINCHAIN_EXE
 The configs_dir (where the config files will reside) can be set through the command line
 or the environment variable RIPPLED_SIDECHAIN_CFG_DIR
 """
+import json
 import os
 import shutil
 import sys
 import traceback
 from pathlib import Path
-from typing import Dict, List, Optional, cast
+from typing import Any, Dict, List, Optional, Union, cast
 
 from jinja2 import Environment, FileSystemLoader
-from xrpl.models import XRP, IssuedCurrency
+from xrpl.models import XRP, Amount, IssuedCurrency
 
 from slk.config.cfg_strs import generate_sidechain_stanza
 from slk.config.config_params import ConfigParams
@@ -76,6 +77,39 @@ def _generate_cfg_dir_mainchain(
     return sub_dir + "/rippled.cfg"
 
 
+def _amt_to_json(amt: Amount) -> Union[str, Dict[str, Any]]:
+    if isinstance(amt, str):
+        return amt
+    else:
+        return amt.to_dict()
+
+
+def _generate_asset_stanzas(assets: Optional[Dict[str, XChainAsset]] = None) -> str:
+    if assets is None:
+        # default to xrp only at a 1:1 value
+        assets = {}
+        assets["xrp_xrp_sidechain_asset"] = XChainAsset(
+            XRP(), XRP(), "1", "1", "400", "400"
+        )
+
+    index_stanza = """
+[sidechain_assets]"""
+
+    asset_stanzas = []
+
+    for name, xchainasset in assets.items():
+        index_stanza += "\n" + name
+        new_stanza = f"""
+[{name}]
+mainchain_asset={json.dumps(_amt_to_json(xchainasset.main_asset))}
+sidechain_asset={json.dumps(_amt_to_json(xchainasset.side_asset))}
+mainchain_refund_penalty={json.dumps(_amt_to_json(xchainasset.main_refund_penalty))}
+sidechain_refund_penalty={json.dumps(_amt_to_json(xchainasset.side_refund_penalty))}"""
+        asset_stanzas.append(new_stanza)
+
+    return index_stanza + "\n" + "\n".join(asset_stanzas)
+
+
 def _generate_validators_txt(sub_dir: str, validators: List[str]) -> None:
     template = JINJA_ENV.get_template("validators.jinja")
 
@@ -116,6 +150,7 @@ def _generate_cfg_dir_sidechain(
         sidenet.federator_keypairs,
         xchain_assets,
     )
+    asset_stanzas = _generate_asset_stanzas(xchain_assets)
 
     fixed_ips_json = [p.to_dict() for p in fixed_ips]
 
@@ -129,6 +164,7 @@ def _generate_cfg_dir_sidechain(
         "mainchain_door_account": sidenet.main_account.classic_address,
         "mainchain_ip": mainnet.url,
         "mainchain_port_ws": mainnet.ports[mainnet_i].ws_public_port,
+        "asset_stanzas": asset_stanzas,
         "sidechain_stanza": sidechain_stanza,
         # other
         "fixed_ips": fixed_ips_json,
